refactor: extract entry modal and helpers
