+++ wisely/lib/widgets/journal_list_item.dart
@@ -1,28 +1,14 @@
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_quill/flutter_quill.dart' as quill;
-import 'package:intl/intl.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
 import 'package:wisely/blocs/audio/player_cubit.dart';
 import 'package:wisely/classes/journal_entities.dart';
-import 'package:wisely/utils/image_utils.dart';
-import 'package:wisely/widgets/editor_widget.dart';
-
-import '../theme.dart';
-import 'audio_player.dart';
-import 'map_widget.dart';
-
-NumberFormat nf = NumberFormat("###.0#", "en_US");
-DateFormat df = DateFormat('yyyy-MM-dd HH:mm:ss');
-
-String formatType(String s) => s.replaceAll('HealthDataType.', '');
-String formatUnit(String s) => s.replaceAll('HealthDataUnit.', '');
-String formatAudio(JournalAudio journalAudio) =>
-    'Audio Note: ${journalAudio.data.duration.toString().split('.')[0]}';
+import 'package:wisely/theme.dart';
+import 'package:wisely/widgets/entry_modal_widget.dart';
+import 'package:wisely/widgets/entry_tools.dart';
 
 class JournalListItem extends StatelessWidget {
   final JournalEntity item;
@@ -94,91 +80,7 @@ class JournalListItem extends StatelessWidget {
                 ),
                 clipBehavior: Clip.antiAliasWithSaveLayer,
                 builder: (BuildContext context) {
-                  return Container(
-                    color: AppColors.entryBgColor,
-                    child: ListView(
-                      shrinkWrap: true,
-                      children: <Widget>[
-                        item.maybeMap(
-                          journalAudio: (audio) => MapWidget(
-                            geolocation: audio.geolocation,
-                          ),
-                          journalImage: (image) => MapWidget(
-                            geolocation: image.geolocation,
-                          ),
-                          journalEntry: (entry) => MapWidget(
-                            geolocation: entry.geolocation,
-                          ),
-                          orElse: () => Container(),
-                        ),
-                        item.maybeMap(
-                          journalAudio: (JournalAudio audio) {
-                            return const AudioPlayerWidget();
-                          },
-                          journalImage: (JournalImage image) {
-                            File file =
-                                File(getFullImagePathWithDocDir(image, docDir));
-                            return Container(
-                              color: Colors.black,
-                              child: Image.file(
-                                file,
-                                cacheHeight: 1200,
-                                width: double.infinity,
-                                height: 400,
-                                fit: BoxFit.scaleDown,
-                              ),
-                            );
-                          },
-                          journalEntry: (JournalEntry journalEntry) {
-                            var editorJson =
-                                json.decode(journalEntry.entryText.quill!);
-
-                            quill.QuillController _controller =
-                                quill.QuillController.basic();
-
-                            _controller = quill.QuillController(
-                                document: quill.Document.fromJson(editorJson),
-                                selection:
-                                    const TextSelection.collapsed(offset: 0));
-
-                            return EditorWidget(
-                              controller: _controller,
-                              height: 240,
-                              readOnly: true,
-                            );
-                          },
-                          quantitative: (qe) => qe.data.map(
-                            cumulativeQuantityData: (qd) => Padding(
-                              padding: const EdgeInsets.all(24.0),
-                              child: InfoText(
-                                'End: ${df.format(qe.meta.dateTo)}'
-                                '\n${formatType(qd.dataType)}: '
-                                '${nf.format(qd.value)} ${formatUnit(qd.unit)}',
-                              ),
-                            ),
-                            discreteQuantityData: (qd) => Padding(
-                              padding: const EdgeInsets.all(24.0),
-                              child: InfoText(
-                                'End: ${df.format(qe.meta.dateTo)}'
-                                '\n${formatType(qd.dataType)}: '
-                                '${nf.format(qd.value)} ${formatUnit(qd.unit)}',
-                              ),
-                            ),
-                          ),
-                          orElse: () => Container(),
-                        ),
-                        Padding(
-                          padding: const EdgeInsets.symmetric(
-                              vertical: 4.0, horizontal: 16.0),
-                          child: InfoText(df.format(item.meta.dateFrom)),
-                        ),
-                        ElevatedButton(
-                          child: const Text('Close'),
-                          onPressed: () => Navigator.pop(context),
-                        )
-                      ],
-                    ),
-                  );
+                  return EntryModalWidget(item: item, docDir: docDir);
                 },
               );
             },
@@ -186,25 +88,5 @@ class JournalListItem extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class InfoText extends StatelessWidget {
-  final String text;
-  final int maxLines;
-  const InfoText(
-    this.text, {
-    Key? key,
-    this.maxLines = 5,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(text,
-        maxLines: maxLines,
-        style: const TextStyle(
-          fontFamily: 'ShareTechMono',
-          fontSize: 14.0,
-        ));
   }
 }
